create memberLeave event and func
